style : 글 없을 경우 레이아웃 스타일 추가
