+++ src/app/(afterLogin)/(main)/daily/[day]/_components/daily/DailyThread.tsx
@@ -6,12 +6,20 @@ import GreyHeartIcon from '@/app/_components/icon/reaction/GreyHeartIcon';
 import { DailyThreadType } from '@/type';
 
 interface Props {
-  data: DailyThreadType;
+  data?: DailyThreadType;
 }
 
 const DUMMY_TARGET_TYPE = 'post';
 
 export default function DailyThread({ data }: Props) {
+  if (!data) {
+    return (
+      <div className="flex flex-col flex-grow justify-center items-center text-center w-full min-h-[15.0rem] font-body text-textDisabled">
+        아직 게시글이 없어요. <br />
+        가장 먼저 글을 남겨보세요.
+      </div>
+    );
+  }
   const { profileUrl, name, comment, reaction, content, postUrl, isEdit } = data;
   return (
     <div className="flex gap-8 w-[37.5rem] min-h-[12.1rem] pl-16 pt-16 pb-[1.4rem] pr-[2.6rem] bg-white border-b border-bgGrey">
